pyproject.toml
[tool.poetry]
name = "experimental"
version = "0.0.1"
description = ""
authors = ["Kodesage Research Team"]
readme = "README.md"
packages = [
    { include = "experimental" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
langchain-core = "^0.3.28"
langchain-community = "^0.3.13"
langchain-openai = "^0.2.14"
langchain-google-genai = "^2.0.6"
langchain-postgres = "^0.0.12"
langchain-ollama = "^0.2.2"
langchain-anthropic = "^0.3.1"
langchain-text-splitters = "^0.3.4"
langgraph = "^0.2.60"
langgraph-checkpoint-postgres = "^2.0.8"
langchain-experimental = "^0.3.3"
langgraph-cli = {extras = ["inmem"], version = "^0.1.65"}
rich = "^13.7.1"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.26.1"
tavily-python= "^0.5.0"

[tool.poetry.scripts]
experimental = "experimental.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
